feat(snacks): add picker for creating a new worktree
I'm not entirely happy how this turned out. It feels wrong to have
a separate picker for creating a new worktree instead of it being an
action of the same picker. But if there is a way to implement that
I couldn't figure it out. The fact that the pickers need different
`finder`, `format`, and `preview` also suggest that these need to be
separate pickers. As does the fact that `git_log`, `git_log_line`, and
`git_log_file` builtin pickers are all separate pickers instead of
a configuration option of the same picker.